asilinks/requirements.txt

numpy==1.15
pandas==0.23.4
celery==4.2.1
# pycurl==7.43

coverage==4.5.1
pyjwt==1.6.4
pillow==5.2.0
requests==2.19.1
weasyprint==0.42.3
python-magic==0.4.15
sentry-sdk==0.4.3

django==2.1
django-nose==1.4.6
django-storages==1.6.6
django-extensions==2.1.3
django-cors-headers==2.4.0
pymongo==3.7.1
mongoengine==0.15.3
git+https://github.com/mongoengine/django-mongoengine.git@v0.3

djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
django-rest-framework-mongoengine==3.3.1
drf-extensions==0.4.0

boto3==1.7.70
pyfcm==1.4.5
paypalrestsdk==1.13.1
